Adapt data types to support larger files / prevent overflow.

+++ src/benchmark/file_io_read_micro_benchmark.hpp
@@ -41,8 +41,8 @@ class FileIOMicroReadBenchmarkFixture : public MicroBenchmarkBasicFixture {
   const ssize_t uint32_t_size = ssize_t{sizeof(uint32_t)};
   std::string filename;
   uint64_t control_sum = uint64_t{0};
-  uint32_t NUMBER_OF_BYTES = uint32_t{0};
-  uint32_t NUMBER_OF_ELEMENTS = uint32_t{0};
+  uint64_t NUMBER_OF_BYTES = uint64_t{0};
+  uint64_t NUMBER_OF_ELEMENTS = uint64_t{0};
   std::vector<uint32_t> numbers = std::vector<uint32_t>{};
   void read_non_atomic_multi_threaded(benchmark::State& state, uint16_t thread_count);
   void read_non_atomic_single_threaded(benchmark::State& state);
